requirements.txt
Flask
Flask-Testing
requests
pytest
locustio